# Conflicts:
#	mlnode/.gitignore
#	mlnode/iotdb/mlnode/client.py
#	mlnode/iotdb/mlnode/config.py
#	mlnode/iotdb/mlnode/constant.py
#	mlnode/iotdb/mlnode/exception.py
#	mlnode/iotdb/mlnode/util.py
#	mlnode/resources/conf/iotdb-mlnode.toml

+++ confignode/src/main/java/org/apache/iotdb/confignode/manager/load/balancer/region/IRegionGroupAllocator.java
@@ -39,7 +39,7 @@ public interface IRegionGroupAllocator {
    */
   TRegionReplicaSet generateOptimalRegionReplicasDistribution(
       Map<Integer, TDataNodeConfiguration> availableDataNodeMap,
-      Map<Integer, Long> freeDiskSpaceMap,
+      Map<Integer, Double> freeDiskSpaceMap,
       List<TRegionReplicaSet> allocatedRegionGroups,
       int replicationFactor,
       TConsensusGroupId consensusGroupId);
