+++ metrics/interface/src/main/java/org/apache/iotdb/metrics/type/AutoGauge.java
@@ -22,8 +22,8 @@ package org.apache.iotdb.metrics.type;
 import java.util.Map;
 
 public interface AutoGauge extends IMetric {
-  /** get value according to the state of obj */
-  long value();
+  /** get value according to the state of obj. */
+  double value();
 
   @Override
   default void constructValueMap(Map<String, Object> result) {
